Agrego el endpoint para actualizar el estado de un anuncio.

+++ src/controllers/anuncioController.ts
@@ -3,11 +3,10 @@ import { Request, Response } from 'express';
 import Anuncio, { IAnuncio } from '../models/Anuncio';
 import Usuario, { IUsuario } from '../models/Usuario';
 import sharp from 'sharp';
-import { BadRequestError, AppError } from '../utils/errors';
+import { BadRequestError, AppError, ForbiddenError } from '../utils/errors';
 import mongoose from 'mongoose';
 import redisClient from '../config/redis';
-import { isOwner } from '../utils/anuncio';
-
+import { EstadosAnuncio, isOwner } from '../utils/anuncio';
 
 // Definir el tipo de respuesta con la población del autor
 interface AnuncioPopulated extends Omit<IAnuncio, 'autor'> {
@@ -26,9 +25,10 @@ interface LeanAnuncio {
     _id: mongoose.Types.ObjectId;
     nombre: string;
     email: string;
-  } | null;
+  };
   fechaPublicacion: Date;
   slug: string;
+  estado: string;
 }
 
 const getAnuncios = async (req: Request, res: Response): Promise<void> => {
@@ -103,6 +103,7 @@ const getAnuncios = async (req: Request, res: Response): Promise<void> => {
         : null,
       fechaPublicacion: anuncio.fechaPublicacion,
       slug: anuncio.slug,
+      estado: anuncio.estado,
     }));
 
     res.status(200).json({
@@ -215,6 +216,7 @@ const getAnunciosUsuario = async (req: Request, res: Response): Promise<void> =>
         : null,
       fechaPublicacion: anuncio.fechaPublicacion,
       slug: anuncio.slug,
+      estado: anuncio.estado,
     }));
 
     res.status(200).json({
@@ -301,9 +303,14 @@ const getAnuncio = async (req: Request, res: Response): Promise<void> => {
 
 const deleteAnuncio = async (req: Request, res: Response): Promise<void> => {
   const { anuncioId } = req.params;
-  
+  const userId = req.userId;
+
   try {
-    const userIsOwner = await isOwner(anuncioId, req.userId);
+    if (!userId) {
+      throw new ForbiddenError();
+    }
+
+    const userIsOwner = await isOwner(anuncioId, userId);
     if (userIsOwner) {
       await Anuncio.deleteOne({ _id: anuncioId });
     }
@@ -325,4 +332,56 @@ const deleteAnuncio = async (req: Request, res: Response): Promise<void> => {
   }
 };
 
-export { LeanAnuncio, getAnuncios, uploadImages, getAnunciosUsuario, getAnuncio, deleteAnuncio };
+const changeStatusAnuncio = async (req: Request, res: Response): Promise<void> => {
+  const { anuncioId } = req.params;
+  const { estado } = req.body;
+  const userId = req.userId;
+  
+  try {
+    if (!userId) {
+      throw new ForbiddenError();
+    }
+
+    const userIsOwner = await isOwner(anuncioId, userId);
+    if (!userIsOwner) {
+      throw new ForbiddenError('Solo el dueño puede actualizar el estado de un anuncio');
+    }
+
+    if (!Object.values(EstadosAnuncio).includes(estado)) {
+      throw new BadRequestError('El estado enviado no existe entre los posibles estados del anuncio');
+    }
+
+    const anuncio = await Anuncio.findOne({ _id: anuncioId });
+    if (!anuncio) {
+      throw new BadRequestError('No existe un anuncio con ese id');
+    }
+
+    if (estado === anuncio.estado) {
+      throw new BadRequestError('El estado enviado es igual al estado actual');
+    }
+
+    if (anuncio.estado === EstadosAnuncio.VENDIDO) {
+      throw new BadRequestError('El anuncio se encuentra en estado vendido, por lo que no puede cambiar de estado');
+    }
+
+    anuncio.estado = estado;
+    anuncio.save();
+    
+    res.status(200).send({ result: 'Estado del anuncio actualizado correctamente' });
+  } catch (error: any) {
+    if (error.status) {
+      res.status(error.status).json({
+        message: error.message,
+        error: true,
+      });
+    } else {
+      res.status(500).json({
+        message: 'Ocurrió un error inesperado',
+        error: true,
+      });
+    }
+  }
+};
+
+
+export { LeanAnuncio, getAnuncios, uploadImages, getAnunciosUsuario, getAnuncio, deleteAnuncio, changeStatusAnuncio };
